Extract createClampsSlice from projectStore (P5 store slice)
